// types/file.ts
export enum FileType {
  IMAGE = "image",
  VIDEO = "video",
  AUDIO = "audio",
  DOCUMENT = "document",
}

export enum ImageMime {
  JPEG = "image/jpeg",
  JPG = "image/jpg",
  PNG = "image/png",
  GIF = "image/gif",
  WEBP = "image/webp",
  SVG = "image/svg+xml",
}

export enum VideoMime {
  MP4 = "video/mp4",
  MPEG = "video/mpeg",
  QUICKTIME = "video/quicktime",
  AVI = "video/x-msvideo",
  MKV = "video/x-matroska",
}

export enum AudioMime {
  MPEG = "audio/mpeg",
  WAV = "audio/wav",
  OGG = "audio/ogg",
  MP4 = "audio/mp4",
}

export enum DocumentMime {
  PDF = "application/pdf",
  DOC = "application/msword",
  DOCX = "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
}